+++ custom_components/iptime_manager/const.py
@@ -9,6 +9,7 @@ PLATFORMS: list[Platform] = [
     Platform.BINARY_SENSOR,
     Platform.SWITCH,
     Platform.SELECT,
+    Platform.NUMBER,
 ]
 
 CONF_URL = 'iptime_url'